fsl_qbman: Fixing of compile issues
Source: https://github.com/Freescale/linux-fslc.git
MR: 107408
Type: Integration
Disposition: Backport from branch 5.4.y+qoriq+fslc
ChangeID: b7738c9a9ff868f26905e56e43946c50764d150d
Description:

pulled the original changes from commit id "5da0f97c1884017b03aa26441058ee1d8e31cf4e"
to fixing of compile issues.

+++ drivers/staging/fsl_qbman/bman_config.c
@@ -281,7 +281,7 @@ static __init int parse_mem_property(struct device_node *node, const char *name,
 	if (!pint || (ret != 16)) {
 		pr_info("No %s property '%s', using memblock_alloc(%016zx)\n",
 				node->full_name, name, *sz);
-		*addr = memblock_alloc(*sz, *sz);
+		*addr = (unsigned long)memblock_alloc(*sz, *sz);
 		vaddr = (unsigned long)phys_to_virt(*addr);
 		if (zero)
 			memset((void *)vaddr, 0, *sz);
